# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Books API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Laravel serve"
        }
    ],
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription utilisateur",
                "description": "Crée un utilisateur et lui accorde un token",
                "operationId": "3443e3fe000081005957217d95b7c7ba",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jean@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "token": "...",
                                    "user": {
                                        "id": 1,
                                        "name": "Jean Dupont",
                                        "email": "jean@mail.com"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur validation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The email field is required.",
                                    "errors": {
                                        "email": [
                                            "The email field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifie un utilisateur. Limité à 10 tentatives par minute",
                "operationId": "1eb51b67697ed1c691a7fbcf25b03450",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "jean@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "token": "...",
                                    "user": {
                                        "id": 1,
                                        "name": "Jean Dupont",
                                        "email": "jean@mail.com"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Identifiants invalides"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "description": "Supprime le token actuel. Il faut être connecté",
                "operationId": "2100e6e17f6eaf1ec83a4b2071aa85ca",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOi..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déconnecté",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Déconnecté"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}